feat: [SECURITY] added abstract class RequestValidator, record RequestedActionRecord, and record RequestActionResponse
RequestValidator is an abstract class which represents the request validation component. It describes the API for calling the component
(through public methods) and, in comments, describes which components it will interact with. Its implementation will handle verifying
an action requested by a user, by first verifying the authentication token provided by the user, as well as verifying that the user
has the proper permissions to carry out the specified action.

RequestedAction is a record which contains the details of a requested action - the user who is requesting the action, as well as the
action they are requesting.

RequestActionResponse is a record which contains the 'response' by the RequestValidator to a request for a RequestedAction - the result
of the action (insufficient permissions, sufficient permissions and successful action, or sufficient permissions but failed action) as
well as any data returned to the user by an action, if carried out successfully.